backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advanced-news-backend"
version = "1.0.0"
description = "Advanced Arabic News System - Backend API"
authors = [{name = "Advanced News Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "Flask>=2.3.3",
    "Flask-CORS>=4.0.0",
    "Flask-SQLAlchemy>=3.0.5",
    "Flask-SocketIO>=5.3.6",
    "python-socketio>=5.8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "feedparser>=6.0.10",
    "openai>=0.28.1",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "redis>=4.6.0",
    "celery>=5.3.1",
    "psycopg2-binary>=2.9.7",
    "gunicorn>=21.2.0",
    "prometheus-client>=0.17.1",
    "APScheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
